requirements.txt
PyPDF2
faiss-cpu
google-generativeai
numpy
